requirements.txt
PyMySQL==1.0.2
python-decouple==3.6
